6-8-2023

+++ app/src/main/java/com/netcast/radio/download/adapter/DownloadEpisodeAdapter.kt
@@ -6,13 +6,15 @@ import com.netcast.radio.base.BaseAdapter
 import com.netcast.radio.databinding.DownloadRowBinding
 import com.netcast.radio.db.AppDatabase
 import com.netcast.radio.ui.radioplayermanager.episodedata.Data
+import kotlinx.coroutines.*
+import java.io.File
 
 class DownloadEpisodeAdapter(
     private var list: MutableList<Data>,
     private val _onClickListenerPodcast: OnClickEpisodeDownload,
     private var isCheckBoxesEnable: Boolean
 ) : BaseAdapter<DownloadRowBinding, Data>(list) {
-    private val selectedItems = mutableListOf<Data>()
+    private var selectedItems = mutableListOf<Data>()
 
     override val layoutId: Int = R.layout.download_row
 
@@ -23,11 +25,15 @@ class DownloadEpisodeAdapter(
             listener = _onClickListenerPodcast
             executePendingBindings()
         }
+  /*      binding.checkboxChecked.setOnCheckedChangeListener { buttonView, isChecked ->
+            currentItem.isSelected = !currentItem.isSelected
+            toggleSelection(currentItem)
+        }*/
+
         binding.checkboxChecked.setOnClickListener {
             currentItem.isSelected = !currentItem.isSelected
             toggleSelection(currentItem)
         }
-
         // Highlight selected items
         binding.checkboxChecked.isChecked = currentItem.isSelected
 
@@ -38,7 +44,7 @@ class DownloadEpisodeAdapter(
     }
 
     override fun getItemsCount(data: List<Data>): Int {
-        return data.size;
+        return data.size
     }
 
     private fun toggleSelection(item: Data) {
@@ -55,18 +61,36 @@ class DownloadEpisodeAdapter(
     }
 
     // Delete selected items
-    fun deleteSelectedItems(appDatabase: AppDatabase?) {
-        for (i in selectedItems)
-            appDatabase?.appDap()?.deleteOfflineEpisodeById(i.id)
+    suspend fun deleteSelectedItems(appDatabase: AppDatabase?) {
+        val job=CoroutineScope(Dispatchers.IO).async {
+            for (i in selectedItems) {
+                var data = appDatabase!!.appDap().getOfflineEpisodeById(i.id)
+                val fileUr = data.fileURI
+                val fdelete: File = File(fileUr)
+                if (fdelete.exists()) {
+                    if (fdelete.delete()) {
+                        appDatabase?.appDap()?.deleteOfflineEpisodeById(i.id)
+                    }
+                }
+
+            }
+        }
+        job.await()
         list.removeAll(selectedItems)
         selectedItems.clear()
+        CoroutineScope(Dispatchers.Main).launch {
         notifyDataSetChanged()
+        }
 
     }
 
     fun enbaleCheckboxes(enable: Boolean) {
         isCheckBoxesEnable = enable
         notifyDataSetChanged()
+    }
+
+    override fun getItemViewType(position: Int): Int {
+        return position
     }
 }
